Fix mangling of zlib API on 64 bit platforms.
It seems https://codereview.chromium.org/8954009/ was a bit
overzealous. On 64 bit platforms, zlib.h attempts to use the 64 bit
API for all callers. But, we were not allowing that, making writing
portable code difficult. This CL updates our manglings in zlib.h so
that on a 64 bit platform the API defaults to the 64 bit version, as
intended.

This change is downstream of https://codereview.chromium.org/13564004/
and must land after it.

R=rsleevi@chromium.org,agl@chromium.org, brettw@chromium.org
BUG=None


Review URL: https://chromiumcodereview.appspot.com/13473020

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 5a4d2679d2e9b672a99aaa4b59cb638a556f2ead