Un-nest configs in GN files.
People sometimes nest targets or configs, usually with the assumption that this limits the visibility of a config to within a target. But this nesting provides no visibility restrictions over declaring it outside of a block.

Un-nest for clarity.

Review URL: https://codereview.chromium.org/1318823008

Cr-Original-Commit-Position: refs/heads/master@{#346461}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 4af2eac8e84692d94f88504ab6e0b244b88dddcb